Refactor validation code structure
